Split columns setting button/content
